Stop comparing createdAt when deciding a request is unsaved
`hasUnsavedChanges` zeroed `modifiedAt` on both sides before comparing,
because it moves on every save, but left `createdAt` in. `newRequest()`
stamps both from `Date.now() / 1000`, so two constructions of the same
content differ by whichever millisecond each was made in -- and a request
rebuilt from defaults rather than loaded, an import or a Postman
round-trip, reports changes nobody made and raises the New Request
confirmation over nothing.

It also made the suite nondeterministic. `workspace.test.ts` builds two
fixtures back to back and expects them to match, which held only while
both landed in the same millisecond; it failed in CI on an unrelated PR
once a run straddled a tick, which is how this was found.

Both timestamps now go through one `content()` helper. Two tests pin it:
one sets the timestamps explicitly, and one waits for a guaranteed clock
tick -- that second one fails every time before this change and passes
every time after, rather than once in a while either way. Checked by
reverting the comparison, which reddens both.

The `hasUnsavedChanges` cases moved to `workspace-unsaved.test.ts`: the
additions took `workspace.test.ts` past the 300-line lint cap, and the
block was self-contained.
